Move grant access overlay for audio inputs
This component will be used across the dock inputs, moving the button
and container to own component will ensure consistency.
